hearing docket fixed
